refactor(catalog): move field-constraints introspection into the engine
Move get_field_constraints + FieldConstraintSchema from apps/claim_edit into
apps/catalog/engine/entity_api/field_constraints/, and repoint the
GET /field-constraints/{entity_type} route in config/api.py. The introspection
derives numeric min/max/step metadata from any ClaimControlledModel's declared
validators — read-only field metadata that belongs with the read surface, not
the interactive write engine.

The two symbols are copied verbatim (bodies byte-identical); the wire URL and
the OpenAPI component name are unchanged, and the now-unused validator /
db_models / Schema imports drop out of claim_write. Their tests move into a new
test_field_constraints.py so the write-engine test file stays focused.

Also give the "Engine internal layers" import contract exhaustive_ignores =
["tests"], so a co-located engine/tests/ directory doesn't fail the exhaustive
layer check — matching the catalog and core contracts (engine needs no apps /
migrations ignore: it has no AppConfig and no migrations of its own).

Also record in the entity-engine design doc a follow-up to make the delete
registrar's delete-preview path caller-injected: it currently hard-codes one
preview schema inside the engine — the lone delete-preview not injected like the
detail / listing / restore serializers — so a domain-named wire shape lives in
the otherwise domain-neutral engine.

No migration; no change to generated frontend types.
